# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Keeper Financial API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/candidate/applications": {
      "get": {
        "tags": [
          "candidate applications"
        ],
        "summary": "List Candidate Applications",
        "operationId": "list_candidate_applications_api_v1_candidate_applications_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-dev-auth-sub",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Auth-Sub"
            }
          },
          {
            "name": "x-dev-auth-aal",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "aal1",
              "title": "X-Dev-Auth-Aal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Candidate access denied"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/candidate/applications/status": {
      "get": {
        "tags": [
          "candidate applications"
        ],
        "summary": "Candidate Status",
        "operationId": "candidate_status_api_v1_candidate_applications_status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-dev-auth-sub",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Auth-Sub"
            }
          },
          {
            "name": "x-dev-auth-aal",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "aal1",
              "title": "X-Dev-Auth-Aal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateStatusListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Candidate access denied"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/candidate/applications/{application_id}": {
      "get": {
        "tags": [
          "candidate applications"
        ],
        "summary": "Get Candidate Application",
        "operationId": "get_candidate_application_api_v1_candidate_applications__application_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          },
          {
            "name": "x-dev-auth-sub",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Auth-Sub"
            }
          },
          {
            "name": "x-dev-auth-aal",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "aal1",
              "title": "X-Dev-Auth-Aal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateApplicationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Candidate access denied"
          },
          "404": {
            "description": "Owned application not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "candidate applications"
        ],
        "summary": "Patch Candidate Application",
        "operationId": "patch_candidate_application_api_v1_candidate_applications__application_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          },
          {
            "name": "x-dev-auth-sub",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Auth-Sub"
            }
          },
          {
            "name": "x-dev-auth-aal",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "aal1",
              "title": "X-Dev-Auth-Aal"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationDraftUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateApplicationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Candidate access denied"
          },
          "404": {
            "description": "Owned application not found"
          },
          "409": {
            "description": "Application revision or lifecycle conflict"
          },
          "422": {
            "description": "Draft validation failed"
          }
        }
      }
    },
    "/api/v1/candidate/applications/{application_id}/submit": {
      "post": {
        "tags": [
          "candidate applications"
        ],
        "summary": "Submit Candidate Application",
        "operationId": "submit_candidate_application_api_v1_candidate_applications__application_id__submit_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          },
          {
            "name": "x-dev-auth-sub",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Auth-Sub"
            }
          },
          {
            "name": "x-dev-auth-aal",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "aal1",
              "title": "X-Dev-Auth-Aal"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationAction"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateApplicationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Candidate access denied"
          },
          "404": {
            "description": "Owned application not found"
          },
          "409": {
            "description": "Application revision or lifecycle conflict"
          },
          "422": {
            "description": "Required submission content is incomplete"
          }
        }
      }
    },
    "/api/v1/candidate/applications/{application_id}/withdraw": {
      "post": {
        "tags": [
          "candidate applications"
        ],
        "summary": "Withdraw Candidate Application",
        "operationId": "withdraw_candidate_application_api_v1_candidate_applications__application_id__withdraw_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          },
          {
            "name": "x-dev-auth-sub",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Auth-Sub"
            }
          },
          {
            "name": "x-dev-auth-aal",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "aal1",
              "title": "X-Dev-Auth-Aal"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationAction"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateApplicationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Candidate access denied"
          },
          "404": {
            "description": "Owned application not found"
          },
          "409": {
            "description": "Application revision or lifecycle conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/candidate/privacy-disclosure": {
      "get": {
        "tags": [
          "candidate applications"
        ],
        "summary": "Candidate Privacy Disclosure",
        "operationId": "candidate_privacy_disclosure_api_v1_candidate_privacy_disclosure_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-dev-auth-sub",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Auth-Sub"
            }
          },
          {
            "name": "x-dev-auth-aal",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "aal1",
              "title": "X-Dev-Auth-Aal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidatePrivacyDisclosureResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Candidate access denied"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplicationAction": {
        "properties": {
          "expected_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_revision"
        ],
        "title": "ApplicationAction"
      },
...
      "ApplicationListResponse": {
        "properties": {
          "applications": {
            "items": {
              "$ref": "#/components/schemas/CandidateApplicationResponse"
            },
            "type": "array",
            "title": "Applications"
          }
        },
        "type": "object",
        "required": [
          "applications"
        ],
        "title": "ApplicationListResponse"
      },
...
      "CandidateApplicationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "recruitment_posting_id": {
            "type": "string",
            "format": "uuid",
            "title": "Recruitment Posting Id"
          },
          "source_posting_slug": {
            "type": "string",
            "title": "Source Posting Slug"
          },
          "source_posting_title": {
            "type": "string",
            "title": "Source Posting Title"
          },
          "source_posting_version": {
            "type": "integer",
            "title": "Source Posting Version"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "state": {
            "type": "string",
            "enum": [
              "draft",
              "submitted",
              "withdrawn"
            ],
            "title": "State"
          },
          "status": {
            "type": "string",
            "enum": [
              "application_started",
              "application_submitted",
              "withdrawn",
              "declined"
            ],
            "title": "Status"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "given_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Given Name"
          },
          "family_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family Name"
          },
          "preferred_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "country_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Code"
          },
          "preferred_contact_method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Contact Method"
          },
          "available_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available From"
          },
          "referral_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referral Source"
          },
          "referral_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referral Detail"
          },
          "interest_statement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interest Statement"
          },
          "relevant_experience": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relevant Experience"
          },
          "employment": {
            "items": {
              "$ref": "#/components/schemas/EmploymentEntryResponse"
            },
            "type": "array",
            "title": "Employment"
          },
          "education": {
            "items": {
              "$ref": "#/components/schemas/EducationEntryResponse"
            },
            "type": "array",
            "title": "Education"
          },
          "privacy_acknowledged": {
            "type": "boolean",
            "title": "Privacy Acknowledged"
          },
          "information_accuracy_confirmed": {
            "type": "boolean",
            "title": "Information Accuracy Confirmed"
          },
          "privacy_disclosure_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Privacy Disclosure Version"
          },
          "privacy_acknowledged_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Privacy Acknowledged At"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "withdrawn_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Withdrawn At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "recruitment_posting_id",
          "source_posting_slug",
          "source_posting_title",
          "source_posting_version",
          "schema_version",
          "revision",
          "state",
          "status",
          "email",
          "given_name",
          "family_name",
          "preferred_name",
          "phone",
          "city",
          "region",
          "country_code",
          "preferred_contact_method",
          "available_from",
          "referral_source",
          "referral_detail",
          "interest_statement",
          "relevant_experience",
          "employment",
          "education",
          "privacy_acknowledged",
          "information_accuracy_confirmed",
          "privacy_disclosure_version",
          "privacy_acknowledged_at",
          "submitted_at",
          "withdrawn_at",
          "created_at",
          "updated_at"
        ],
        "title": "CandidateApplicationResponse"
      },
...
      "CandidatePrivacyDisclosureResponse": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "paragraphs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Paragraphs"
          }
        },
        "type": "object",
        "required": [
          "title",
          "version",
          "paragraphs"
        ],
        "title": "CandidatePrivacyDisclosureResponse"
      },
      "CandidateStatusListResponse": {
        "properties": {
          "applications": {
            "items": {
              "$ref": "#/components/schemas/CandidateVisibleApplicationStatus"
            },
            "type": "array",
            "title": "Applications"
          }
        },
        "type": "object",
        "required": [
          "applications"
        ],
        "title": "CandidateStatusListResponse"
      },
      "CandidateVisibleApplicationStatus": {
        "properties": {
          "application_id": {
            "type": "string",
            "format": "uuid",
            "title": "Application Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "application_started",
              "application_submitted",
              "withdrawn",
              "declined"
            ],
            "title": "Status"
          },
          "messages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Messages"
          }
        },
        "type": "object",
        "required": [
          "application_id",
          "status",
          "messages"
        ],
        "title": "CandidateVisibleApplicationStatus"
      },
...
      "EducationEntryResponse": {
        "properties": {
          "institution_name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Institution Name"
          },
          "program_name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Program Name"
          },
          "completion_year": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1900.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Completion Year"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "institution_name",
          "program_name"
        ],
        "title": "EducationEntryResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}